Authentication Implemented (#2)

+++ src/main/java/org/kainos/ea/resources/EmployeeController.java
@@ -1,12 +1,16 @@
 package org.kainos.ea.resources;
 
 import io.swagger.annotations.Api;
+import org.kainos.ea.api.AuthService;
 import org.kainos.ea.api.EmployeeService;
 import org.kainos.ea.cli.DeliveryEmployeeRequest;
+import org.kainos.ea.cli.UserRole;
+import org.kainos.ea.client.AuthenticationException;
 import org.kainos.ea.cli.DeliveryEmployeeUpdateRequest;
 import org.kainos.ea.client.GenericActionFailedException;
 import org.kainos.ea.client.GenericValidationException;
 import org.kainos.ea.client.GenericDoesNotExistException;
+import org.kainos.ea.db.AuthDao;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
@@ -16,12 +20,17 @@ import javax.ws.rs.core.Response;
 @Path("/api")
 public class EmployeeController {
     private final EmployeeService employeeService = new EmployeeService();
+    private final AuthService authService = new AuthService(new AuthDao());
 
     @GET
     @Path("/employees")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getAllOrders() {
+    public Response getAllOrders(@QueryParam("token") String token) {
         try {
+            if(!authService.doesTokenHaveRole(token,UserRole.HR)){
+                return Response.status(Response.Status.FORBIDDEN).build();
+            }
+
             return Response
                     .status(Response.Status.OK)
                     .entity(employeeService.getAllEmployees())
@@ -29,14 +38,21 @@ public class EmployeeController {
         } catch (GenericActionFailedException e) {
             System.err.println(e.getMessage());
             return Response.serverError().build();
+        } catch (AuthenticationException e) {
+            System.err.println(e.getMessage());
+            return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
         }
     }
 
     @POST
     @Path("/employees/delivery")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response createOrder(DeliveryEmployeeRequest deliveryEmployee) {
+    public Response createOrder(DeliveryEmployeeRequest deliveryEmployee, @QueryParam("token") String token) {
         try {
+            if(!authService.doesTokenHaveRole(token,UserRole.HR)){
+                return Response.status(Response.Status.FORBIDDEN).build();
+            }
+
             return Response
                     .ok(employeeService.createDeliveryEmployee(deliveryEmployee))
                     .build();
@@ -49,14 +65,21 @@ public class EmployeeController {
                     .status(Response.Status.BAD_REQUEST)
                     .entity(e.getMessage())
                     .build();
+        } catch (AuthenticationException e) {
+            System.err.println(e.getMessage());
+            return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
         }
     }
 
     @GET
     @Path("/employees/delivery/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getDeliveryEmployeeById(@PathParam("id") int id) {
+    public Response getDeliveryEmployeeById(@PathParam("id") int id,@QueryParam("token") String token) {
         try {
+            if(!authService.doesTokenHaveRole(token,UserRole.HR)){
+                return Response.status(Response.Status.FORBIDDEN).build();
+            }
+
             return Response
                     .status(Response.Status.OK)
                     .entity(employeeService.getDeliveryEmployeeById(id))
@@ -65,8 +88,11 @@ public class EmployeeController {
             System.err.println(e.getMessage());
             return Response.serverError().build();
         } catch (GenericDoesNotExistException e){
-
+            System.err.println(e.getMessage());
             return Response.status(Response.Status.NOT_FOUND).entity(e.getMessage()).build();
+        } catch (AuthenticationException e) {
+            System.err.println(e.getMessage());
+            return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
         }
     }
 
@@ -74,8 +100,12 @@ public class EmployeeController {
     @GET
     @Path("/employees/delivery")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getAllDeliveryEmployees(){
+    public Response getAllDeliveryEmployees(@QueryParam("token") String token){
         try {
+            if(!authService.doesTokenHaveRole(token,UserRole.HR)){
+                return Response.status(Response.Status.FORBIDDEN).build();
+            }
+
             return Response
                     .status(Response.Status.OK)
                     .entity(employeeService.getAllDeliveryEmployees())
@@ -83,15 +113,21 @@ public class EmployeeController {
         } catch (GenericActionFailedException e) {
             System.err.println(e.getMessage());
             return Response.serverError().build();
+        } catch (AuthenticationException e) {
+            System.err.println(e.getMessage());
+            return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
         }
     }
 
     @DELETE
     @Path("/employees/delivery/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response deleteDeliveryEmployee(@PathParam("id") int id) {
-
+    public Response deleteDeliveryEmployee(@PathParam("id") int id, @QueryParam("token") String token) {
         try {
+            if(!authService.doesTokenHaveRole(token,UserRole.HR)){
+                return Response.status(Response.Status.FORBIDDEN).build();
+            }
+
             employeeService.deleteDeliveryEmployee(id);
 
             return Response.ok().build();
@@ -100,7 +136,37 @@ public class EmployeeController {
             System.err.println(e.getMessage());
             return Response.serverError().build();
         } catch (GenericDoesNotExistException e) {
+            System.err.println(e.getMessage());
             return Response.status(Response.Status.NOT_FOUND).entity(e.getMessage()).build();
+        } catch (AuthenticationException e) {
+            System.err.println(e.getMessage());
+            return Response.status(Response.Status.BAD_REQUEST).entity(e.getMessage()).build();
+        }
+    }
+
+
+    @PUT
+    @Path("/employees/delivery/{id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response createEmployee(@PathParam("id") int id, DeliveryEmployeeRequest employee,@QueryParam("token") String token) {
+        try{
+            if(!authService.doesTokenHaveRole(token,UserRole.HR)){
+                return Response.status(Response.Status.FORBIDDEN).build();
+            }
+
+            employeeService.updateEmployee(id, employee);
+
+            return Response.ok().build();
+        } catch (GenericActionFailedException e) {
+            System.err.println(e.getMessage());
+            return Response.serverError().build();
+
+        } catch (GenericValidationException | GenericDoesNotExistException | AuthenticationException e) {
+            System.err.println(e.getMessage());
+            return Response
+                    .status(Response.Status.BAD_REQUEST)
+                    .entity(e.getMessage())
+                    .build();
         }
     }
